requirements.txt
flask==3.0.3
gunicorn==22.0.0
openai==1.52.0
gspread==6.1.2
requests==2.32.3
Pillow==10.4.0
